+++ packages/utils/src/index.ts
@@ -5,6 +5,7 @@ import createErrorHandler from './createErrorHandler';
 import createSchemaUtils from './createSchemaUtils';
 import dataURItoBlob from './dataURItoBlob';
 import dateRangeOptions from './dateRangeOptions';
+import deepClone from './deepClone';
 import deepEquals from './deepEquals';
 import englishStringTranslator from './englishStringTranslator';
 import enumOptionsDeselectValue from './enumOptionsDeselectValue';
@@ -118,6 +119,7 @@ export {
   DateElement,
   dataURItoBlob,
   dateRangeOptions,
+  deepClone,
   deepEquals,
   descriptionId,
   englishStringTranslator,